converted int types to NS(U)Integers using ConvertCocoa64 script

+++ NSArray+BSJSONAdditions.h
@@ -29,6 +29,6 @@
 
 + (NSArray *)arrayWithJSONString:(NSString *)jsonString;
 - (NSString *)jsonStringValue;
-- (NSString *)jsonStringValueWithIndentLevel:(int)level;
+- (NSString *)jsonStringValueWithIndentLevel:(NSInteger)level;
 
 @end
